Day 9, 1 star
Very naive implementation, want to move to using `FileBlocks` that are themselves the list of points vs a long list, but I cannot seem to get the swap operations to play nice.

When I move to the more advanced model using the sample input everything works fine. On the puzzle input, it breaks down.